🐛 fix: tighten cellular check + add diagnostic logging (#202)
User report: on real-device fresh install, picker tap landed on
`ModelDownloadView.downloadingView` (progress bar) instead of the demo
replay host body. Progress bar advanced normally so DL was happening,
but the demo never appeared. The only state machine path that produces
this exact symptom is `isCellularNow()` returning `true` despite the
user being on Wi-Fi — which causes `initialLoad` to bail before
loading demo sources, leaving `sources = []` and `replayVM = nil`
forever.

Most likely cause: `path.isExpensive` was the cellular check, and it
also returns `true` for personal hotspot, metered Wi-Fi, VPN-managed
paths, and (in some configurations) iCloud Private Relay. So real
Wi-Fi can be flagged expensive when the user expects "Wi-Fi means
demo plays."

Tighten `isCellularNow` to use `path.usesInterfaceType(.cellular)`
instead — only literal cellular interfaces trip the safety net. The
hotspot / metered-Wi-Fi coverage that `isExpensive` provided is
deferred to #191's full cellular-consent modal, which will gate
`ModelManager.startDownload` properly.

Also add `os.Logger` diagnostic notices throughout `initialLoad` and
`isCellularNow`. Subsystem `com.tyabu12.Pastura` category
`DemoReplayHostView` — visible in Console.app on a Mac connected to
the device. If the issue recurs, the log line will name exactly which
exit branch fired and the path's cellular / wifi / expensive flags.

Move `isCellularNow` and the shared `logger` static to
`DemoReplayHostView+Routing.swift` to keep the host view's main file
under the 400-line `file_length` cap.
